requirements.txt
# NOVA-AI Discord Bot Requirements

# Discord library
discord.py>=2.3.2

# Google Generative AI
google-generativeai>=0.3.0

# Environment variable management
python-dotenv>=1.0.0
